pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "career-page-crawler"
version = "0.1.0"
description = "Keyword-driven job posting crawler using JobSpy + LLM extraction"
requires-python = ">=3.11"
dependencies = [
    "python-jobspy>=1.1.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "playwright>=1.40",
    "playwright-stealth>=1.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "langchain>=0.3",
    "langgraph>=0.2",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
bkng = [
    "bkng-ml[rs]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
crawler = "crawler.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[[tool.pip.index]]
url = "https://jfrog.booking.com/artifactory/api/pypi/pypi/simple"
name = "booking-artifactory"

[tool.ruff]
line-length = 120
target-version = "py311"
